requirements/base.txt
asgiref
certifi
cffi
charset-normalizer
cmake
cryptography
defusedxml
Django
django-allauth
django-bootstrap5
django-ckeditor
django-js-asset
django_debug_toolbar
djangorestframework
idna
oauthlib
Pillow
pycparser
PyJWT
PyMySQL
python3-openid
pytz
requests
requests-oauthlib
sqlparse
typing_extensions
urllib3
wincertstore